requirements.txt
anyio==4.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.2
idna==3.6
lxml==4.9.3
numpy==1.26.2
outcome==1.3.0.post0
pandas==2.1.4
polygon==1.1.3
polygon-api-client==1.13.3
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
restclient==0.11.0
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wsproto==1.2.0
xmltodict==0.12.0
xmltojson==2.0.2
